Update MainActivity.java
In the getGameOverMessage the totalQuestion return size of whole ArrayList that is 13, however the player chooses only 6 question to answer in the game. 

To reflect this I added -7 to the totalQuestion, so it shows sth like: "You got 3 right out of 6", instead of "You got 3 right out of 13".

Couldn't figure out better solution, but knowing the size of ArrayList I just subtract so it would return 6.